Add new sub-module for BumpTransactionEvent
Its accompanying event handler will also live here.
